backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-orjson-renderer==1.7.1
orjson==3.8.3
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
